Add files via upload
changed all the doubles to int in calculatespeed return types

+++ Car/Parts/Aerodynamics.cpp
@@ -19,7 +19,7 @@ void Aerodynamics::removePart(string part) {
     else DepartmentOutput::removePart(part);
 }
 
-double Aerodynamics::getSpeed(){
+int Aerodynamics::getSpeed(){
     return speed * 1000;
 }
